% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_arima.R
\name{auto_arima}
\alias{auto_arima}
\title{Auto ARIMA por Minimização do AIC}
\usage{
auto_arima(y, pmax = 6, qmax = 6, dmax = 2)
}
\arguments{
\item{y}{A série real para se estimar os coeficientes.}

\item{pmax}{a maior ordem do ar a ser estimado para procurar o AIC mínimo.}

\item{qmax}{a maior ordem do ma a ser estimado para procurar o AIC mínimo.}

\item{dmax}{a maior ordem da diferença a ser estimada para procurar o AIC mínimo.}
}
\description{
Encontra as parâmetros p, q e d do modelo ARIMA que minimizem a estatística AIC.
}
\examples{
y = simular_serie(
  dist="rnorm",
  args=list(n=100, mean=0, sd=1),
  H0=list(intercepto=0,ar=c(0.5, 0.25),ma=c())
)
auto_arima(y, pmax=3, qmax=3, dmax=1)

}
