% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Minimização de AIC.R
\name{min.AIC}
\alias{min.AIC}
\title{min.AIC}
\usage{
\method{min}{AIC}(Y, pmax = 6, qmax = 6, dmax = 2)
}
\arguments{
\item{Y}{A série real para se estimar os coeficientes.}

\item{pmax}{a maior ordem do ar a ser estimado para procurar o AIC mínimo.}

\item{qmax}{a maior ordem do ma a ser estimado para procurar o AIC mínimo.}

\item{dmax}{a maior ordem da diferença a ser estimada para procurar o AIC mínimo.}
}
\description{
min.AIC
}
