% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rony.R
\name{rony}
\alias{rony}
\title{Rony}
\usage{
rony(Y, p, q, intercept = TRUE)
}
\arguments{
\item{Y}{Uma série de tempo.}

\item{p}{A ordem do ar a ser estimado.}

\item{q}{A ordem do ma a ser estimado.}
}
\description{
Rony
}
