% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sazonalidade.R
\name{sazonalidade}
\alias{sazonalidade}
\title{Estimação de sazonalidade}
\usage{
sazonalidade(y, sazon)
}
\arguments{
\item{y}{Um vetor numérico com a série de tempo sem tendência a ser modelada.}

\item{sazon}{Uma matriz com as dummies de sazonalidade. Sem nenhuma sazonalidade por padrão.}
}
\description{
Estima a sazonalidade de uma série de tempo y de acordo com opções de formato de sazonalidade, sem intercepto.
}
\examples{
#sazonalidade(y = c(1,2,1,3,1,4,1,2,3,1,2,4,1,2,3), sazon = c(1,0,0,1,0,0,1,0,0,1,0,0,1,0,0)

}
