% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simular_serie.R
\name{simular_serie}
\alias{simular_serie}
\title{Simulação de Série}
\usage{
simular_serie(dist, args, H0)
}
\arguments{
\item{dist}{A ser passado para função erros. O comportamento do erro do modelo. O nome (string) de qualqer função geradora de do pacote {stats} <https://stat.ethz.ch/R-manual/R-devel/library/stats/html/Distributions.html>, ou "arma" para modelar o erro como um arma (o erro desse modelo é uma normal padrão).}

\item{args}{A ser passado para função erros. Uma lista com os parâmetros que a função escolhida em dist precisem. Se escolhido arma, passar uma lista no formato: list(p.erros = ..vetor 1xp com os primeiros p erros.., dist = ..distribuição dos erros dentre as opções de stats.., args = ..argumentos da distribuição.., intercepto = ..valor do intercepto.., ar = ..vetor com os coeficientes ar.., ma = ..vetor com os coeficientes ma..).}

\item{H0}{Uma lista com a hipótese nula sobre os parâmetros intercepto, ar e ma passados em model. Formato: list(intercepto = ...., ar=c(....), ma=c(....))}
}
\description{
Recebe os parâmetros de um modelo ARMA e o formato do erro e simula uma série de tempo.
}
\examples{
simular_serie(
  dist = "rnorm",
  args = list(n=20, mean=0, sd=1),
  H0 = list(intercepto = 0, ar = c(0.3,0.2,0.1), ma = c(0.4,0.2))
)

}
