% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tendencia.R
\name{tendencia}
\alias{tendencia}
\title{Estimação de tendência}
\usage{
tendencia(y, trend)
}
\arguments{
\item{y}{Um vetor numérico com a série de tempo a ser modelada.}

\item{trend}{Uma string com o tipo de tendência a ser estimada: "poli.n" para polinômio de ordem n, "expo.n" para exponencial de ordem n, "hpfl.n" para filtro hp de sensibilidade n. Sem nenhuma tendência por padrão.}
}
\description{
Estima a tendência de uma série de tempo y de acordo com opções de formato de tendência.
}
\examples{
#tendencia(y = c(1,2,1,3,1,4,1,2,3,1,2,4,1,2,3), trend = "hpfl.1600")

}
