% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Teste t generalizado.R
\name{test.raiz.unit}
\alias{test.raiz.unit}
\title{Teste t generalizado}
\usage{
test.raiz.unit(
  Y,
  dist,
  args,
  H0,
  S,
  a = 0.05,
  trend = FALSE,
  sazon = FALSE,
  cicle = FALSE
)
}
\arguments{
\item{Y}{A série real para se estimar os coeficientes.}

\item{dist}{A ser passado para função erros. O comportamento do erro do modelo. O nome (string) de qualqer função geradora de do pacote {stats} <https://stat.ethz.ch/R-manual/R-devel/library/stats/html/Distributions.html>, ou "arma" para modelar o erro como um arma (o erro desse modelo é uma normal padrão) [EXPANDIR PARAR OUTRAS POSSIBILIDADES].}

\item{args}{A ser passado para função erros. Uma lista com os parâmetros que a função escolhida em dist precisem. Se escolhido arma, passar uma lista no formato: list(p.erros = ..vetor 1xp com os primeiros p erros.., dist = ..distribuição dos erros dentre as opções de stats.., args = ..argumentos da distribuição.., intercepto = ..valor do intercepto.., ar = ..vetor com os coeficientes ar.., ma = ..vetor com os coeficientes ma..).}

\item{H0}{Uma lista com a hipótese nula sobre os parâmetros intercepto, ar e ma passados em model.
Formato: list(intercepto = ...., ar=c(....), ma=c(....))}

\item{S}{Número de simulações a serem feitas para obter a distribuição empírica de Y}

\item{a}{O valor do nível de significãncia (quantil desejado).}

\item{trend}{A ser passado para função tendencia. Uma string com o tipo de tendência a ser estimada: "poli.n" para polinômio de ordem n, "exp.n" para exponencial de ordem n, "hp.n" para filtro hp de sensibilidade n. Sem nenhuma tendência por padrão.}

\item{sazon}{A ser passado para função sazonalidade.  matriz com as dummies de sazonalidade [EXPANDIR PARA MODELOS DE SAZONALIDADE]. Sem nenhuma sazonalidade por padrão.}

\item{cicle}{Um ciclo já estimado para se controlado por.}
}
\description{
Realizar um teste t nos coeficientes de um modelo ARMA(p,q), sensível à tendência e sazonalidade, para distribuições de estatística não-normais padrão
}
