% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erros.R
\name{erros}
\alias{erros}
\title{Gerador de erros}
\usage{
erros(dist, args)
}
\arguments{
\item{dist}{O comportamento do erro do modelo. O nome (string) de qualqer função geradora de do pacote {stats} <https://stat.ethz.ch/R-manual/R-devel/library/stats/html/Distributions.html>, ou "arma" para modelar o erro como um arma (o erro desse modelo é uma normal padrão).}

\item{args}{Uma lista com os parâmetros que a função escolhida em dist precisem. Se escolhido arma, passar uma lista no formato: list(dist = ..distribuição dos erros dentre as opções de stats.., args = ..argumentos da distribuição.., intercepto = ..valor do intercepto.., ar = ..vetor com os coeficientes ar.., ma = ..vetor com os coeficientes ma..).}
}
\description{
Cria um vetor de erros aleatórios, criados de acordo com a distribuição escolhida e seus parâmetros.
}
\examples{
erros("rnorm", args=list(n=10, mean=0, sd=1))
erros("arma", args=list(
  dist="rnorm",
  args=list(n=20, mean=1, sd=1),
  intercepto=0,
  ar=c(0.3, 0.2, 0.1),
  ma=c(0.4, 0.2))
)

}
