% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimação de sazonalidade.R
\name{sazonalidade}
\alias{sazonalidade}
\title{Estimação de sazonalidade}
\usage{
sazonalidade(Y, sazon)
}
\arguments{
\item{Y}{Um vetor numérico com a série de tempo sem tendência a ser modelada.}

\item{sazon}{Uma matriz com as dummies de sazonalidade [EXPANDIR PARA MODELOS DE SAZONALIDADE]. Sem nenhuma sazonalidade por padrão.}
}
\description{
Estima a sazonalidade de uma série de tempo Y de acordo com opções de formato de sazonalidade, sem intercepto.
}
