% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimação de tendência.R
\name{tendencia}
\alias{tendencia}
\title{Estimação de tendência}
\usage{
tendencia(Y, trend)
}
\arguments{
\item{Y}{Um vetor numérico com a série de tempo a ser modelada.}

\item{trend}{Uma string com o tipo de tendência a ser estimada: "poli.n" para polinômio de ordem n, "expo.n" para exponencial de ordem n, "hpfl.n" para filtro hp de sensibilidade n. Sem nenhuma tendência por padrão.}
}
\description{
Estima a tendência de uma série de tempo Y de acordo com opções de formato de tendência.
}
